Reorganized GUI code blocks into smaller chunks
There also seems to be a problem with threading the TOF graph data
generation, so it's left to run in the main thread (in the function
process_redraw_tof_data()).

+++ Lib/fermi_libraries/fermi_libraries/common_functions.py
@@ -673,7 +673,7 @@ def lrange(start, end, step=1, remove=None):
     return changed_list
 
 def closest(locs, array):
-    indices = np.array([np.argmin((array - loc)**2) for loc in locs])
+    indices = np.array([np.argmin((array - loc)**2) for loc in locs], dtype=int)
     return indices
 
 def non_normalized_gaussians(params, x):
